refactor(analysis): move complexity unit tests
